skrting

+++ conv1_ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: ecross <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 12:47:51 by ecross            #+#    #+#             */
-/*   Updated: 2019/11/14 12:42:58 by ecross           ###   ########.fr       */
+/*   Updated: 2019/11/14 16:53:26 by ecross           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ char	*u_convert(va_list arg_list, t_list *list)
 	char	*var;
 
 	prec = list->flag_vals[get_pos(list->flag_chars, '.')];
-	arg = va_arg(arg_list, unsigned int);
+	arg = va_arg(arg_list, int);
 	if (arg == 0 && prec == 0
 			&& list->flag_found[get_pos(list->flag_chars, '.')] == 1)
 		return (malloc_empty_string());
@@ -93,7 +93,7 @@ char	*u_convert(va_list arg_list, t_list *list)
 	else if (list->h == 1)
 		var = ft_itoa((long long unsigned)((short unsigned)arg));
 	else
-		var = ft_u_itoa((long long unsigned)arg);
+		var = ft_u_itoa((long long unsigned)(unsigned int)arg);
 	i = 0;
 	while (var[i])
 		i++;
