+++ conv2_ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: ecross <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/12 15:34:19 by ecross            #+#    #+#             */
-/*   Updated: 2019/11/12 15:34:35 by ecross           ###   ########.fr       */
+/*   Updated: 2019/11/14 16:56:39 by ecross           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,13 +20,13 @@ char	*x_convert(va_list arg_list, t_list *list)
 	char	*var;
 
 	prec = list->flag_vals[get_pos(list->flag_chars, '.')];
-	arg = va_arg(arg_list, unsigned int);
+	arg = va_arg(arg_list, int);
 	if (list->h == 2)
 		var = hex_convert((long long unsigned)((char unsigned)arg), 87);
 	else if (list->h == 1)
 		var = hex_convert((long long unsigned)((short unsigned)arg), 87);
 	else
-		var = hex_convert((long long unsigned)arg, 87);
+		var = hex_convert((long long unsigned)(unsigned int)arg, 87);
 	i = 0;
 	while (var[i])
 		i++;
@@ -43,13 +43,13 @@ char	*xx_convert(va_list arg_list, t_list *list)
 	char	*var;
 
 	prec = list->flag_vals[get_pos(list->flag_chars, '.')];
-	arg = va_arg(arg_list, unsigned int);
+	arg = va_arg(arg_list, int);
 	if (list->h == 2)
 		var = hex_convert((long long unsigned)((char unsigned)arg), 55);
 	else if (list->h == 1)
 		var = hex_convert((long long unsigned)((short unsigned)arg), 55);
 	else
-		var = hex_convert((long long unsigned)arg, 55);
+		var = hex_convert((long long unsigned)(unsigned int)arg, 55);
 	i = 0;
 	while (var[i])
 		i++;
